# app/MaPSCDAService/wwwroot/swagger/v1/maps-cda.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MaPSCDAService, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://maps-api-management-dev.azure-api.net/maps-cda-service/"
    }
  ],
  "paths": {
    "/claims-gathering-redirect": {
      "get": {
        "tags": [
          "ClaimsRedirect"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/RequestHeaderModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/rqp": {
      "get": {
        "tags": [
          "MapsCdaService"
        ],
        "parameters": [
          {
            "name": "headerModel",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/RequestHeaderModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/redirect-details": {
      "post": {
        "tags": [
          "MapsCdaService"
        ],
        "parameters": [
          {
            "name": "headerModel",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/RequestHeaderModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/csrf-token": {
      "get": {
        "summary": "Get CSRF token",
        "description": "Returns a CSRF token and sets it as an HttpOnly cookie.",
        "responses": {
          "200": {
            "description": "CSRF token set",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RequestHeaderModel": {
        "required": [
          "iss",
          "userSessionId"
        ],
        "type": "object",
        "properties": {
          "userSessionId": {
            "minLength": 1,
            "type": "string"
          },
          "iss": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "correlationId": {
            "maxLength": 36,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
  "tags": [
    {
      "name": "ClaimsRedirect"
    },
    {
      "name": "MapsCdaService"
    }
  ]
}